# BusinessProject/bin/Debug/net7.0/OpenAPIs/swagger15.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "MSJogos",
    "version": "1.0"
  },
  "paths": {
...
    "/updatejogos": {
      "put": {
        "tags": [
          "AdminJogos"
        ],
        "parameters": [
          {
            "name": "ligaID",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/updateequipas": {
      "put": {
        "tags": [
          "AdminJogos"
        ],
        "parameters": [
          {
            "name": "ligaID",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/topscorers": {
      "put": {
        "tags": [
          "AdminJogos"
        ],
        "parameters": [
          {
            "name": "IdLiga",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "get": {
        "tags": [
          "Jogos"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ResponseTopScorers"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ResponseTopScorers"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ResponseTopScorers"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
...
    "/GetLogsJogos": {
      "get": {
        "tags": [
          "AdminJogos"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LogJogos"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LogJogos"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LogJogos"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/GetLogsEquipas": {
      "get": {
        "tags": [
          "AdminJogos"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LogEquipas"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LogEquipas"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LogEquipas"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BirthTopScorers": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "nullable": true
          },
          "place": {
            "type": "string",
            "nullable": true
          },
          "country": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CardsTopScorers": {
        "type": "object",
        "properties": {
          "yellow": {
            "type": "integer",
            "format": "int32"
          },
          "yellowred": {
            "type": "integer",
            "format": "int32"
          },
          "red": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "DribblesTopScorers": {
        "type": "object",
        "properties": {
          "attempts": {
            "type": "integer",
            "format": "int32"
          },
          "success": {
            "type": "integer",
            "format": "int32"
          },
          "past": {
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "DuelsTopScorers": {
        "type": "object",
        "properties": {
          "total": {
            "type": "integer",
            "format": "int32"
          },
          "won": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "FoulsTopScorers": {
        "type": "object",
        "properties": {
          "drawn": {
            "type": "integer",
            "format": "int32"
          },
          "committed": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "GamesTopScorers": {
        "type": "object",
        "properties": {
          "appearences": {
            "type": "integer",
            "format": "int32"
          },
          "lineups": {
            "type": "integer",
            "format": "int32"
          },
          "minutes": {
            "type": "integer",
            "format": "int32"
          },
          "number": {
            "nullable": true
          },
          "position": {
            "type": "string",
            "nullable": true
          },
          "rating": {
            "type": "string",
            "nullable": true
          },
          "captain": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "GoalsTopScorers": {
        "type": "object",
        "properties": {
          "total": {
            "type": "integer",
            "format": "int32"
          },
          "conceded": {
            "type": "integer",
            "format": "int32"
          },
          "assists": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "saves": {
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "LeagueTopScorers": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "country": {
            "type": "string",
            "nullable": true
          },
          "logo": {
            "type": "string",
            "nullable": true
          },
          "flag": {
            "type": "string",
            "nullable": true
          },
          "season": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "LogEquipas": {
        "type": "object",
        "properties": {
          "dataCria": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "equipa": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "LogJogos": {
        "type": "object",
        "properties": {
          "dataCria": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "idJogo": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "PassesTopScorers": {
        "type": "object",
        "properties": {
          "total": {
            "type": "integer",
            "format": "int32"
          },
          "key": {
            "type": "integer",
            "format": "int32"
          },
          "accuracy": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "PenaltyTopScorers": {
        "type": "object",
        "properties": {
          "won": {
            "nullable": true
          },
          "commited": {
            "nullable": true
          },
          "scored": {
            "type": "integer",
            "format": "int32"
          },
          "missed": {
            "type": "integer",
            "format": "int32"
          },
          "saved": {
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "PlayerTopScorers": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "firstname": {
            "type": "string",
            "nullable": true
          },
          "lastname": {
            "type": "string",
            "nullable": true
          },
          "age": {
            "type": "integer",
            "format": "int32"
          },
          "birth": {
            "$ref": "#/components/schemas/BirthTopScorers"
          },
          "nationality": {
            "type": "string",
            "nullable": true
          },
          "height": {
            "type": "string",
            "nullable": true
          },
          "weight": {
            "type": "string",
            "nullable": true
          },
          "injured": {
            "type": "boolean"
          },
          "photo": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ResponseTopScorers": {
        "type": "object",
        "properties": {
          "player": {
            "$ref": "#/components/schemas/PlayerTopScorers"
          },
          "statistics": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StatisticTopScorers"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ShotsTopScorers": {
        "type": "object",
        "properties": {
          "total": {
            "type": "integer",
            "format": "int32"
          },
          "on": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "StatisticTopScorers": {
        "type": "object",
        "properties": {
          "team": {
            "$ref": "#/components/schemas/TeamTopScorers"
          },
          "league": {
            "$ref": "#/components/schemas/LeagueTopScorers"
          },
          "games": {
            "$ref": "#/components/schemas/GamesTopScorers"
          },
          "substitutes": {
            "$ref": "#/components/schemas/SubstitutesTopScorers"
          },
          "shots": {
            "$ref": "#/components/schemas/ShotsTopScorers"
          },
          "goals": {
            "$ref": "#/components/schemas/GoalsTopScorers"
          },
          "passes": {
            "$ref": "#/components/schemas/PassesTopScorers"
          },
          "tackles": {
            "$ref": "#/components/schemas/TacklesTopScorers"
          },
          "duels": {
            "$ref": "#/components/schemas/DuelsTopScorers"
          },
          "dribbles": {
            "$ref": "#/components/schemas/DribblesTopScorers"
          },
          "fouls": {
            "$ref": "#/components/schemas/FoulsTopScorers"
          },
          "cards": {
            "$ref": "#/components/schemas/CardsTopScorers"
          },
          "penalty": {
            "$ref": "#/components/schemas/PenaltyTopScorers"
          }
        },
        "additionalProperties": false
      },
...
      "SubstitutesTopScorers": {
        "type": "object",
        "properties": {
          "in": {
            "type": "integer",
            "format": "int32"
          },
          "out": {
            "type": "integer",
            "format": "int32"
          },
          "bench": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "TacklesTopScorers": {
        "type": "object",
        "properties": {
          "total": {
            "type": "integer",
            "format": "int32"
          },
          "blocks": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "interceptions": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "TeamTopScorers": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "logo": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}